Extracted LoadPath into its own module
